More errors for incorrect arrays, AutoExpanded logic, removed schema.py
Consolidated schema.py into metadata.py
