+++ redfish_service_validator/metadata.py
@@ -5,9 +5,16 @@
 import os
 import time
 from collections import Counter, OrderedDict, defaultdict
-import redfish_service_validator.schema as schema
+from collections import namedtuple
+from bs4 import BeautifulSoup
+from functools import lru_cache
+import os.path
 
-# TODO: Use catalog.py instead of schema.py, remove uses of schema.py
+from redfish_service_validator.helper import getNamespace, getNamespaceUnversioned
+
+import logging
+my_logger = logging.getLogger(__name__)
+
 
 EDM_NAMESPACE = "http://docs.oasis-open.org/odata/ns/edm"
 EDMX_NAMESPACE = "http://docs.oasis-open.org/odata/ns/edmx"
@@ -15,6 +22,7 @@ EDM_TAGS = ['Action', 'Annotation', 'Collection', 'ComplexType', 'EntityContaine
             'Member', 'NavigationProperty', 'Parameter', 'Property', 'PropertyRef', 'PropertyValue', 'Record',
             'Schema', 'Singleton', 'Term', 'TypeDefinition']
 EDMX_TAGS = ['DataServices', 'Edmx', 'Include', 'Reference']
+
 
 def bad_edm_tags(tag):
     return tag.namespace == EDM_NAMESPACE and tag.name not in EDM_TAGS
@@ -92,10 +100,8 @@ class Metadata(object):
         self.elapsed_secs = time.time() - start
         self.schema_obj = None
         if data:
-            soup = schema.BeautifulSoup(data, "xml")
-            self.schema_obj = schema.rfSchema(soup, '$metadata', 'service')
-            self.md_soup = self.schema_obj.soup
-            self.service_refs = self.schema_obj.refs
+            self.md_soup = BeautifulSoup(data, "xml")
+            self.service_refs = getReferenceDetails(self.md_soup)
             self.success_get = True
             # set of namespaces included in $metadata
             self.metadata_namespaces = {k for k in self.service_refs.keys()}
@@ -121,11 +127,10 @@ class Metadata(object):
             logger.debug('Metadata: bad_schema_uris = {}'.format(self.bad_schema_uris))
             logger.debug('Metadata: bad_namespace_include = {}'.format(self.bad_namespace_include))
             for ref in self.service_refs:
+                print(ref)
                 name, uri = self.service_refs[ref]
-                self.schema_store[name] = schema.getSchemaObject(self.service, name, uri)
-                if self.schema_store[name] is not None:
-                    for ref in self.schema_store[name].refs:
-                        pass
+                success, soup, origin = getSchemaDetails(service, name, uri)
+                self.schema_store[name] = soup
         else:
             logger.warning('Metadata: getSchemaDetails() did not return success')
 
@@ -192,7 +197,7 @@ class Metadata(object):
             if '#' in schema_uri:
                 schema_uri, frag = k.split('#', 1)
             schema_type = os.path.basename(os.path.normpath(k)).strip('.xml').strip('_v1')
-            success, soup, _ = schema.getSchemaDetails(self.service, schema_type, schema_uri)
+            success, soup, _ = getSchemaDetails(self.service, schema_type, schema_uri)
             if success:
                 for namespace in self.uri_to_namespaces[k]:
                     if soup.find('Schema', attrs={'Namespace': namespace}) is None:
@@ -309,3 +314,197 @@ class OrderedCounter(Counter, OrderedDict):
 
     def __reduce__(self):
         return self.__class__, (OrderedDict(self),)
+
+def storeSchemaToLocal(xml_data, origin, service):
+    """storeSchemaToLocal
+
+    Moves data pulled from service/online to local schema storage
+
+    Does NOT do so if preferonline is specified
+
+    :param xml_data: data being transferred
+    :param origin: origin of xml pulled
+    """
+    config = service.config
+    SchemaLocation = config['metadatafilepath']
+    if not os.path.isdir(SchemaLocation):
+        os.makedirs(SchemaLocation)
+    if 'localFile' not in origin and '$metadata' not in origin:
+        __, xml_name = origin.rsplit('/', 1)
+        new_file = os.path.join(SchemaLocation, xml_name)
+        if not os.path.isfile(new_file):
+            with open(new_file, "w") as filehandle:
+                filehandle.write(xml_data)
+                my_logger.info('Writing online XML to file: {}'.format(xml_name))
+        else:
+            my_logger.info('NOT writing online XML to file: {}'.format(xml_name))
+
+
+@lru_cache(maxsize=64)
+def getSchemaDetails(service, SchemaType, SchemaURI):
+    """
+    Find Schema file for given Namespace.
+
+    param SchemaType: Schema Namespace, such as ServiceRoot
+    param SchemaURI: uri to grab schema, given LocalOnly is False
+    return: (success boolean, a Soup object, origin)
+    """
+    my_logger.debug('getting Schema of {} {}'.format(SchemaType, SchemaURI))
+
+    if SchemaType is None:
+        return False, None, None
+
+    if service is None:
+        return getSchemaDetailsLocal(SchemaType, SchemaURI, {})
+
+    elif service.active and getNamespace(SchemaType) in service.metadata.schema_store:
+        result = service.metadata.schema_store[getNamespace(SchemaType)]
+        if result is not None:
+            return True, result.soup, result.origin
+
+    success, soup, origin = getSchemaDetailsLocal(SchemaType, SchemaURI, service.config)
+    if success:
+        return success, soup, origin
+
+    xml_suffix = '_v1.xml'
+
+    if (SchemaURI is not None) or (SchemaURI is not None and '/redfish/v1/$metadata' in SchemaURI):
+        # Get our expected Schema file here
+        # if success, generate Soup, then check for frags to parse
+        #   start by parsing references, then check for the refLink
+        if '#' in SchemaURI:
+            base_schema_uri, frag = tuple(SchemaURI.rsplit('#', 1))
+        else:
+            base_schema_uri, frag = SchemaURI, None
+        success, data, response, elapsed = service.callResourceURI(base_schema_uri)
+        if success:
+            soup = BeautifulSoup(data, "xml")
+            # if frag, look inside xml for real target as a reference
+            if frag is not None:
+                # prefer type over frag, truncated down
+                # using frag, check references
+                frag = getNamespace(SchemaType)
+                frag = frag.split('.', 1)[0]
+                refType, refLink = getReferenceDetails(
+                    soup, name=base_schema_uri).get(frag, (None, None))
+                if refLink is not None:
+                    success, linksoup, newlink = getSchemaDetails(service, refType, refLink)
+                    if success:
+                        return True, linksoup, newlink
+                    else:
+                        my_logger.error(
+                            "SchemaURI couldn't call reference link {} inside {}".format(frag, base_schema_uri))
+                else:
+                    my_logger.error(
+                        "SchemaURI missing reference link {} inside {}".format(frag, base_schema_uri))
+                    # error reported; assume likely schema uri to allow continued validation
+                    uri = 'http://redfish.dmtf.org/schemas/v1/{}{}'.format(frag, xml_suffix)
+                    my_logger.info("Continue assuming schema URI for {} is {}".format(SchemaType, uri))
+                    return getSchemaDetails(service, SchemaType, uri)
+            else:
+                storeSchemaToLocal(data, base_schema_uri, service)
+                return True, soup, base_schema_uri
+        else:
+            my_logger.debug("SchemaURI called unsuccessfully: {}".format(base_schema_uri))
+    return getSchemaDetailsLocal(SchemaType, SchemaURI, service.config)
+
+
+def getSchemaDetailsLocal(SchemaType, SchemaURI, config):
+    """
+    Find Schema file for given Namespace, from local directory
+
+    param SchemaType: Schema Namespace, such as ServiceRoot
+    param SchemaURI: uri to grab schem (generate information from it)
+    return: (success boolean, a Soup object, origin)
+    """
+    Alias = getNamespaceUnversioned(SchemaType)
+    SchemaLocation, SchemaSuffix = config['metadatafilepath'], '_v1.xml'
+    if SchemaURI is not None:
+        uriparse = SchemaURI.split('/')[-1].split('#')
+        xml = uriparse[0]
+    else:
+        my_logger.warning("SchemaURI was empty, must generate xml name from type {}".format(SchemaType)),
+        return getSchemaDetailsLocal(SchemaType, Alias + SchemaSuffix, config)
+    my_logger.debug(('local', SchemaType, SchemaURI, SchemaLocation + '/' + xml))
+    filestring = Alias + SchemaSuffix if xml is None else xml
+    try:
+        # get file
+        with open(SchemaLocation + '/' + xml, "r") as filehandle:
+            data = filehandle.read()
+
+        # get tags
+        soup = BeautifulSoup(data, "xml")
+        edmxTag = soup.find('Edmx', recursive=False)
+        parentTag = edmxTag.find('DataServices', recursive=False)
+        child = parentTag.find('Schema', recursive=False)
+        SchemaNamespace = child['Namespace']
+        FoundAlias = SchemaNamespace.split(".")[0]
+        my_logger.debug(FoundAlias)
+
+        if FoundAlias in Alias:
+            return True, soup, "localFile:" + SchemaLocation + '/' + filestring
+
+    except FileNotFoundError:
+        # if we're looking for $metadata locally... ditch looking for it, go straight to file
+        if '/redfish/v1/$metadata' in SchemaURI and Alias != '$metadata':
+            my_logger.debug("Unable to find a xml of {} at {}, defaulting to {}".format(SchemaURI, SchemaLocation, Alias + SchemaSuffix))
+            return getSchemaDetailsLocal(SchemaType, Alias + SchemaSuffix, config)
+        else:
+            my_logger.warning("Schema file {} not found in {}".format(filestring, SchemaLocation))
+            if Alias == '$metadata':
+                my_logger.warning("If $metadata cannot be found, Annotations may be unverifiable")
+    except Exception as ex:
+        my_logger.error("A problem when getting a local schema has occurred {}".format(SchemaURI))
+        my_logger.warning("output: ", exc_info=True)
+    return False, None, None
+
+
+def check_redfish_extensions_alias(name, namespace, alias):
+    """
+    Check that edmx:Include for Namespace RedfishExtensions has the expected 'Redfish' Alias attribute
+    :param name: the name of the resource
+    :param item: the edmx:Include item for RedfishExtensions
+    :return: bool
+    """
+    if alias is None or alias != 'Redfish':
+        msg = ("In the resource {}, the {} namespace must have an alias of 'Redfish'. The alias is {}. " +
+               "This may cause properties of the form [PropertyName]@Redfish.TermName to be unrecognized.")
+        my_logger.error(msg.format(name, namespace,
+                             'missing' if alias is None else "'" + str(alias) + "'"))
+        return False
+    return True
+
+
+def getReferenceDetails(soup, metadata_dict=None, name='xml'):
+    """
+    Create a reference dictionary from a soup file
+
+    param arg1: soup
+    param metadata_dict: dictionary of service metadata, compare with
+    return: dictionary
+    """
+    includeTuple = namedtuple('include', ['Namespace', 'Uri'])
+    refDict = {}
+
+    maintag = soup.find("Edmx", recursive=False)
+    reftags = maintag.find_all('Reference', recursive=False)
+    for ref in reftags:
+        includes = ref.find_all('Include', recursive=False)
+        for item in includes:
+            uri = ref.get('Uri')
+            ns, alias = (item.get(x) for x in ['Namespace', 'Alias'])
+            if ns is None or uri is None:
+                my_logger.error("Reference incorrect for: {}".format(item))
+                continue
+            if alias is None:
+                alias = ns
+            refDict[alias] = includeTuple(ns, uri)
+            # Check for proper Alias for RedfishExtensions
+            if name == '$metadata' and ns.startswith('RedfishExtensions.'):
+                check_bool = check_redfish_extensions_alias(name, ns, alias)
+
+    cntref = len(refDict)
+    if metadata_dict is not None:
+        refDict.update(metadata_dict)
+    my_logger.debug("METADATA: References generated from {}: {} out of {}".format(name, cntref, len(refDict)))
+    return refDict
